Moved audio settings to its own class
