+++ Assets/Scripts/Settings/UISettingsSound.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 85575792b20387a409179c22b616f5f2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
